Added timer to delete databases

+++ llistat/views.py
@@ -11,6 +11,18 @@ import json
 from plotly.utils import PlotlyJSONEncoder
 import plotly.express as px
 import os
+import threading
+import time
+
+
+def delete_objects_later(minutes=5):
+    """
+    Deletes objects from models after 5 minutes (300 seconds).
+    """
+    seconds = minutes * 60
+    time.sleep(seconds)
+    CSVData.objects.all().delete()
+    Report.objects.all().delete()
 
 
 @login_required(login_url='llistat:login')
@@ -56,6 +68,9 @@ def upload_csv(request):
                             == "false" else None
                         )
 
+                    if "Usuario" in df.columns:
+                        df = df.drop(columns=["Usuario"])
+
                 except Exception as e:
                     if os.path.exists(csv_path):
                         os.remove(csv_path)
@@ -88,7 +103,6 @@ def upload_csv(request):
                             Cerrado = row["Cerrado"],
                             Fecha_Cierre = row["Fecha_Cierre"],
                             Tarea = row["Tarea"],
-                            Usuario = row["Usuario"],
                             FECHA_I = row["FECHA_I"],
                             Fecha_Fin_Prevista = row["Fecha_Fin_Prevista"],
                             FECHA_F = row["FECHA_F"],
@@ -120,6 +134,7 @@ def upload_csv(request):
                         'error': f"Error saving CSV data to database: {e}"
                     })
 
+
                 return redirect('llistat:report')
 
         else:
@@ -138,48 +153,57 @@ def upload_csv(request):
 @login_required(login_url='llistat:login')
 def report(request):
 
-    Report.objects.all().delete()
+    queryset = CSVData.objects.all()
+    if queryset.count() == 0:
+        return redirect('llistat:upload_csv')
+    else:
+        Report.objects.all().delete()
 
-    OT_sections = ["Mechanical Design", "New solutions"]
-    MRP_sections = ["Purchasing"]
-    hide_costumers = ["TECNOC"]
+        OT_sections = ["Mechanical Design", "New solutions"]
+        MRP_sections = ["Purchasing"]
+        hide_costumers = ["TECNOC"]
 
-    queryset = CSVData.without_F.all()
-    data = list(queryset.values())
-    df = pd.DataFrame(data)
+        queryset = CSVData.without_F.all()
+        data = list(queryset.values())
+        df = pd.DataFrame(data)
 
-    df = df.loc[~df["Peticionario"].isin(hide_costumers)] # Hide not used costumers
-    OT_OMs_series = df.loc[df["Seccion"].isin(OT_sections), "OM"] # Filter OMs with section OT
-    df_final = df.loc[df["Seccion"].isin(MRP_sections), ["Fecha_Fin_Prevista", "OM", "Peticionario", "FECHA_ENTREGA",
-                                                         "FECHA_PREVISTA_MRP"]]
+        df = df.loc[~df["Peticionario"].isin(hide_costumers)] # Hide not used costumers
+        OT_OMs_series = df.loc[df["Seccion"].isin(OT_sections), "OM"] # Filter OMs with section OT
+        df_final = df.loc[df["Seccion"].isin(MRP_sections), ["Fecha_Fin_Prevista", "OM", "Peticionario", "FECHA_ENTREGA",
+                                                             "FECHA_PREVISTA_MRP"]]
 
-    def check_OMs(OM: pd.Series) -> str:
-        if OM in OT_OMs_series.values:
-            return "OT"
-        else:
-            return "MRP"
+        def check_OMs(OM: pd.Series) -> str:
+            if OM in OT_OMs_series.values:
+                return "OT"
+            else:
+                return "MRP"
 
-    df_final["Seccion"] = df_final["OM"].apply(check_OMs)
-    df_final.sort_values(by=["Fecha_Fin_Prevista", "FECHA_ENTREGA"], inplace=True)
+        df_final["Seccion"] = df_final["OM"].apply(check_OMs)
+        df_final.sort_values(by=["Fecha_Fin_Prevista", "FECHA_ENTREGA"], inplace=True)
 
-    records = [
-        Report(
-            Fecha_Fin_Prevista = row["Fecha_Fin_Prevista"],
-            OM = row["OM"],
-            Peticionario = row["Peticionario"],
-            FECHA_ENTREGA = row["FECHA_ENTREGA"],
-            FECHA_PREVISTA_MRP = row["FECHA_PREVISTA_MRP"],
-            Seccion = row["Seccion"]
-        )
-        for _, row in df_final.iterrows()
-    ]
-    Report.objects.bulk_create(records)
+        records = [
+            Report(
+                Fecha_Fin_Prevista = row["Fecha_Fin_Prevista"],
+                OM = row["OM"],
+                Peticionario = row["Peticionario"],
+                FECHA_ENTREGA = row["FECHA_ENTREGA"],
+                FECHA_PREVISTA_MRP = row["FECHA_PREVISTA_MRP"],
+                Seccion = row["Seccion"]
+            )
+            for _, row in df_final.iterrows()
+        ]
+        Report.objects.bulk_create(records)
 
-    report = Report.objects.all()
-    for item in report:
-        item.is_past_date = (item.Fecha_Fin_Prevista and item.Fecha_Fin_Prevista < date.today())
-        item.is_delay = (item.FECHA_ENTREGA and item.FECHA_ENTREGA < date.today())
-    return render(request, 'llistat/report.html', {'report': report})
+        report = Report.objects.all()
+
+        for item in report:
+            item.is_past_date = (item.Fecha_Fin_Prevista and item.Fecha_Fin_Prevista < date.today())
+            item.is_delay = (item.FECHA_ENTREGA and item.FECHA_ENTREGA < date.today())
+
+        thread = threading.Thread(target=delete_objects_later)
+        thread.start()
+
+        return render(request, 'llistat/report.html', {'report': report})
 
 
 def user_login(request):
@@ -205,70 +229,75 @@ def user_login(request):
                    'error_message': error_message})
 @login_required(login_url='llistat:login')
 def plot_report(request):
-    report = Report.objects.all()
-    data = list(report.values())
-    df = pd.DataFrame(data)
-    df['Fecha_Fin_Prevista'] = pd.to_datetime(df['Fecha_Fin_Prevista'], yearfirst=True)
 
-    fig1 = px.pie(df['Seccion'].value_counts(),
-                 values='count',
-                 names=df['Seccion'].value_counts().index,
-                 template="plotly_white",
-                 height=400, width=400,
-                 title="% Secciones por tarea total")
-    fig_json_1 = json.dumps(fig1, cls=PlotlyJSONEncoder)
+    queryset = Report.objects.all()
+    if queryset.count() == 0:
+        return redirect('llistat:upload_csv')
+    else:
+        report = Report.objects.all()
+        data = list(report.values())
+        df = pd.DataFrame(data)
+        df['Fecha_Fin_Prevista'] = pd.to_datetime(df['Fecha_Fin_Prevista'], yearfirst=True)
 
-    report_pet, report_pet_delay = costumers_delay(df)
-    stacked_data = pd.DataFrame({
-        "Peticionarios": report_pet.index,  # x-axis
-        "En curso": report_pet['count'],  # First series
-        "Retrasada": report_pet_delay['count']  # Second series
-    })
-    stacked_data['Peticionarios'] = stacked_data.index  # Correct column values to match original index
+        fig1 = px.pie(df['Seccion'].value_counts(),
+                     values='count',
+                     names=df['Seccion'].value_counts().index,
+                     template="plotly_white",
+                     height=400, width=400,
+                     title="% Secciones por tarea total")
+        fig_json_1 = json.dumps(fig1, cls=PlotlyJSONEncoder)
 
-    fig2 = px.bar(
-        stacked_data.sort_values(by='En curso', ascending=False),
-        x="Peticionarios",  # x-axis
-        y=["En curso", "Retrasada"],  # Stack these series
-        labels={
-            "Peticionarios": "Peticionarios",  # x-axis label
-            "value": "Total tareas",  # y-axis label
-            "variable": "Estado tarea"  # Label for legend
-        },
-        template="plotly_white",
-        height=400,
-        width=750,
-        title="Top 10 Peticionarios"
-    )
-    fig_json_2 = json.dumps(fig2, cls=PlotlyJSONEncoder)
+        report_pet, report_pet_delay = costumers_delay(df)
+        stacked_data = pd.DataFrame({
+            "Peticionarios": report_pet.index,  # x-axis
+            "En curso": report_pet['count'],  # First series
+            "Retrasada": report_pet_delay['count']  # Second series
+        })
+        stacked_data['Peticionarios'] = stacked_data.index  # Correct column values to match original index
 
-    today = datetime.today()
-    report_delay_seccion = df.loc[(df['Fecha_Fin_Prevista'] < today)]['Seccion'].value_counts()
+        fig2 = px.bar(
+            stacked_data.sort_values(by='En curso', ascending=False),
+            x="Peticionarios",  # x-axis
+            y=["En curso", "Retrasada"],  # Stack these series
+            labels={
+                "Peticionarios": "Peticionarios",  # x-axis label
+                "value": "Total tareas",  # y-axis label
+                "variable": "Estado tarea"  # Label for legend
+            },
+            template="plotly_white",
+            height=400,
+            width=750,
+            title="Top 10 Peticionarios"
+        )
+        fig_json_2 = json.dumps(fig2, cls=PlotlyJSONEncoder)
 
-    fig3 = px.pie(report_delay_seccion,
-                 values='count',
-                 names=report_delay_seccion.index,
-                 template="plotly_white",
-                 height=400, width=400,
-                 title='% Sección por tareas en atraso')
+        today = datetime.today()
+        report_delay_seccion = df.loc[(df['Fecha_Fin_Prevista'] < today)]['Seccion'].value_counts()
 
-    fig_json_3 = json.dumps(fig3, cls=PlotlyJSONEncoder)
+        fig3 = px.pie(report_delay_seccion,
+                     values='count',
+                     names=report_delay_seccion.index,
+                     template="plotly_white",
+                     height=400, width=400,
+                     title='% Sección por tareas en atraso')
 
-    df['delay'] = df['Fecha_Fin_Prevista'].apply(lambda x: 'Atraso' if x < today else 'En curso')
-    agg_data = df.groupby(['Seccion', 'delay']).size().reset_index(name='count')
+        fig_json_3 = json.dumps(fig3, cls=PlotlyJSONEncoder)
 
-    fig4 = px.sunburst(data_frame=agg_data,
-                       values='count',
-                       path=['Seccion', 'delay'],
-                       color='delay',
-                       color_discrete_map={'Atraso': 'red', 'En curso': 'green', 'MRP': 'blue', 'OT': 'orange'},
-                       template="plotly_white",
-                       height=400, width=750,
-                       title="Tareas por sección y estado")
+        df['delay'] = df['Fecha_Fin_Prevista'].apply(lambda x: 'Atraso' if x < today else 'En curso')
+        agg_data = df.groupby(['Seccion', 'delay']).size().reset_index(name='count')
 
-    fig_json_4 = json.dumps(fig4, cls=PlotlyJSONEncoder)
+        fig4 = px.sunburst(data_frame=agg_data,
+                           values='count',
+                           path=['Seccion', 'delay'],
+                           color='delay',
+                           color_discrete_map={'Atraso': 'red', 'En curso': 'green', 'MRP': 'blue', 'OT': 'orange'},
+                           template="plotly_white",
+                           height=400, width=750,
+                           title="Tareas por sección y estado")
 
-    return render(request, 'llistat/plots.html', {'fig_json_1': fig_json_1,
-                                                                     'fig_json_2': fig_json_2,
-                                                                     'fig_json_3': fig_json_3,
-                                                                     'fig_json_4': fig_json_4})
+        fig_json_4 = json.dumps(fig4, cls=PlotlyJSONEncoder)
+
+        return render(request, 'llistat/plots.html', {'fig_json_1': fig_json_1,
+                                                                         'fig_json_2': fig_json_2,
+                                                                         'fig_json_3': fig_json_3,
+                                                                         'fig_json_4': fig_json_4})
